Refactor: 앱 패키지 생성
