+++ 1Week_Mission/src/main/java/com/example/mission/app/base/entity/BaseEntity.java
@@ -1,0 +1,36 @@
+package com.example.mission.app.base.entity;
+
+import static javax.persistence.GenerationType.*;
+import static lombok.AccessLevel.*;
+
+import java.time.LocalDateTime;
+
+import javax.persistence.EntityListeners;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+import javax.persistence.MappedSuperclass;
+
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
+import lombok.experimental.SuperBuilder;
+
+@Getter
+@SuperBuilder
+@MappedSuperclass
+@NoArgsConstructor(access = PROTECTED)
+@EntityListeners(AuditingEntityListener.class)
+@ToString
+public class BaseEntity {
+	@Id
+	@GeneratedValue(strategy = IDENTITY)
+	private Long id;
+	@CreatedDate
+	private LocalDateTime createDate;
+	@LastModifiedDate
+	private LocalDateTime modifyDate;
+}
