+++ 1Week_Mission/src/main/java/com/example/mission/app/member/exception/AlreadyExistException.java
@@ -1,0 +1,4 @@
+package com.example.mission.app.member.exception;
+
+public class AlreadyExistException extends RuntimeException {
+}
