+++ 1Week_Mission/src/main/java/com/example/mission/app/post/dto/PostModifyDto.java
@@ -1,0 +1,18 @@
+package com.example.mission.app.post.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class PostModifyDto {
+	private Long id;
+	private String subject;
+	private String content;
+}
