+++ 1Week_Mission/src/main/java/com/example/mission/app/post/repository/PostRepository.java
@@ -1,0 +1,8 @@
+package com.example.mission.app.post.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.example.mission.app.post.entity.Post;
+
+public interface PostRepository extends JpaRepository<Post, Long> {
+}
