+++ 1Week_Mission/src/main/java/com/example/mission/app/product/entity/Product.java
@@ -1,0 +1,22 @@
+package com.example.mission.app.product.entity;
+
+import javax.persistence.Entity;
+
+import com.example.mission.app.base.entity.BaseEntity;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
+import lombok.experimental.SuperBuilder;
+
+@Entity
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor
+@SuperBuilder
+@ToString(callSuper = true)
+public class Product extends BaseEntity {
+	private String subject;
+	private String price;
+}
