--- 1Week_Mission/src/main/java/com/example/mission/member/exception/AlreadyExistException.java
@@ -1,4 +1,0 @@
-package com.example.mission.member.exception;
-
-public class AlreadyExistException extends RuntimeException {
-}
