--- 1Week_Mission/src/main/java/com/example/mission/post/dto/WriteDto.java
@@ -1,21 +1,0 @@
-package com.example.mission.post.dto;
-
-import javax.validation.constraints.NotEmpty;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class WriteDto {
-	@NotEmpty
-	private String subject;
-	@NotEmpty
-	private String content;
-}
