--- 1Week_Mission/src/main/java/com/example/mission/post/entity/PostHashTag.java
@@ -1,23 +1,0 @@
-package com.example.mission.post.entity;
-
-import javax.persistence.Entity;
-
-import com.example.mission.base.entity.BaseEntity;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.ToString;
-import lombok.experimental.SuperBuilder;
-
-@Entity
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-@SuperBuilder
-@ToString(callSuper = true)
-public class PostHashTag extends BaseEntity {
-	private Long memberId;
-	private Long postId;
-	private Long postKeywordId;
-}
