--- 1Week_Mission/src/main/java/com/example/mission/post/repository/PostRepository.java
@@ -1,8 +1,0 @@
-package com.example.mission.post.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.example.mission.post.entity.Post;
-
-public interface PostRepository extends JpaRepository<Post, Long> {
-}
