--- 1Week_Mission/src/main/java/com/example/mission/product/entity/Product.java
@@ -1,22 +1,0 @@
-package com.example.mission.product.entity;
-
-import javax.persistence.Entity;
-
-import com.example.mission.base.entity.BaseEntity;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.ToString;
-import lombok.experimental.SuperBuilder;
-
-@Entity
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-@SuperBuilder
-@ToString(callSuper = true)
-public class Product extends BaseEntity {
-	private String subject;
-	private String price;
-}
