--- 1Week_Mission/src/main/java/com/example/mission/security/SecurityConfig.java
@@ -1,58 +1,0 @@
-package com.example.mission.security;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
-import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.header.writers.frameoptions.XFrameOptionsHeaderWriter;
-import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
-
-import com.example.mission.security.service.MemberSecurityService;
-
-import lombok.RequiredArgsConstructor;
-
-@Configuration
-@EnableWebSecurity
-@EnableGlobalMethodSecurity(prePostEnabled = true)
-@RequiredArgsConstructor
-public class SecurityConfig {
-
-	private final MemberSecurityService memberSecurityService;
-
-	@Bean
-	public SecurityFilterChain securityFilterChain(
-		HttpSecurity http, AuthenticationConfiguration authenticationConfiguration) throws Exception {
-		http.authorizeRequests().antMatchers("/**").permitAll()
-			.and()
-			.headers()
-			.addHeaderWriter(new XFrameOptionsHeaderWriter(
-				XFrameOptionsHeaderWriter.XFrameOptionsMode.SAMEORIGIN))
-			.and()
-			.csrf().disable()
-			.formLogin()
-			.loginPage("/member/login")
-			.defaultSuccessUrl("/")
-			.and()
-			.logout()
-			.logoutRequestMatcher(new AntPathRequestMatcher("/member/logout"))
-			.logoutSuccessUrl("/");
-
-		return http.build();
-	}
-
-	@Bean
-	public PasswordEncoder passwordEncoder() {
-		return new BCryptPasswordEncoder();
-	}
-
-	@Bean
-	public AuthenticationManager authenticationManager(AuthenticationConfiguration authenticationConfiguration) throws Exception {
-		return authenticationConfiguration.getAuthenticationManager();
-	}
-}
